transformations/spellchecker_perturbation/requirements.txt
pyenchant==3.2.0
